Chore: implemented loading state for dashboard
